requirements.txt
aerich==0.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
aiosqlite==0.21.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
async-stripe==6.1.0
async-timeout==5.0.1
asyncclick==8.1.8.0
asyncpg==0.30.0
attrs==25.1.0
bcrypt==4.0.1
bidict==0.23.1
billiard==4.2.1
black==25.1.0
boto3==1.37.37
botocore==1.37.37
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
coloredlogs==15.0.1
cryptography==44.0.2
dictdiffer==0.9.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.115.11
fastapi-storages==0.3.0
firebase-admin==6.8.0
frozenlist==1.7.0
google-api-core==2.25.0rc1
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.1.1
grpcio==1.73.0
grpcio-status==1.73.0
h11==0.16.0
httplib2==0.22.0
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
iso8601==2.1.0
izulu==0.5.4
Jinja2==3.1.6
jmespath==1.0.1
kombu==5.4.2
Mako==1.3.9
MarkupSafe==3.0.2
msgpack==1.1.1
multidict==6.1.0
mypy-extensions==1.0.0
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.21.1
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psycopg2-binary==2.9.10
pyasn1==0.4.8
pyasn1_modules==0.4.1
pycparser==2.22
pycron==3.1.2
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
PyJWT==2.10.1
pyparsing==3.2.3
pypika-tortoise==0.6.1
pytest==8.3.5
pytest-asyncio==0.25.3
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-jose==3.4.0
python-multipart==0.0.20
python-socketio==5.12.1
pytz==2025.2
redis==5.2.1
requests==2.32.4
rsa==4.9.1
s3transfer==0.11.5
sentry-sdk==2.26.1
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
stripe==6.1.0
taskiq==0.11.14
taskiq-dependencies==1.5.7
taskiq-fastapi==0.3.4
taskiq-redis==1.0.3
tomli==2.2.1
tornado==6.5.1
tortoise-orm==0.25.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.34.3
vine==5.1.0
wsproto==1.2.0
yarl==1.18.3
zipp==3.21.0
